+++ include/shadermovebar.h
@@ -33,13 +33,13 @@ public:
 
 	void setLinkVisuToBox(LinkVisuToBox *linkVisuToBox);
 
-	float proxyWidth() const;
+	qreal proxyWidth() const;
 
-	float proxyHeight() const;
+	qreal proxyHeight() const;
 
-	void setProxyHeight(float proxyHeight);
+	void setProxyHeight(qreal proxyHeight);
 
-	void setProxyWidth(float proxyWidth);
+	void setProxyWidth(qreal proxyWidth);
 
 	QVariant itemChange(QGraphicsItem::GraphicsItemChange change, const QVariant &value);
 
@@ -53,8 +53,8 @@ private:
 	QString m_title;
 	int m_fontSize = 12;
 	LinkVisuToBox *m_linkVisuToBox;
-	float m_proxyWidth;
-	float m_proxyHeight;
+	qreal m_proxyWidth;
+	qreal m_proxyHeight;
 };
 
 #endif // SHADERMOVEBAR_H
